requirements.txt
beautifulsoup4==4.13.4
google-api-python-client==2.169.0
google-cloud-storage==3.1.0
google-cloud-vision==3.10.1
gspread==6.2.1
selenium==4.32.0
requests~=2.32.3
protobuf~=6.31.0
google-auth-oauthlib~=1.2.2
pydantic~=2.11.5
pillow~=11.2.1
tldextract~=5.3.0
datamodel-code-generator~=0.30.1
webdriver-manager~=4.0.2